+++ src/main/java/xyz/nikitacartes/easyauth/config/ExtendedConfigV1.java
@@ -95,6 +95,9 @@ public class ExtendedConfigV1 extends ConfigTemplate {
     @ConfigSerializable
     public static final class Aliases {
 
+        Aliases() {
+        }
+
         Aliases(boolean login, boolean register) {
             this.login = login;
             this.register = register;