+++ src/main/java/xyz/nikitacartes/easyauth/config/MainConfigV1.java
@@ -14,6 +14,7 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 @ConfigSerializable
 public class MainConfigV1 extends ConfigTemplate {
     public boolean premiumAutoLogin = true;
+    public boolean offlineByDefault = false;
     public boolean floodgateAutoLogin = true;
     public long maxLoginTries = 3;
     public long kickTimeout = 60;
@@ -43,6 +44,7 @@ public class MainConfigV1 extends ConfigTemplate {
     protected String handleTemplate() throws IOException {
         Map<String, String> configValues = new HashMap<>();
         configValues.put("premiumAutologin", wrapIfNecessary(premiumAutoLogin));
+        configValues.put("offlineByDefault", wrapIfNecessary(offlineByDefault));
         configValues.put("floodgateAutologin", wrapIfNecessary(floodgateAutoLogin));
         configValues.put("maxLoginTries", wrapIfNecessary(maxLoginTries));
         configValues.put("kickTimeout", wrapIfNecessary(kickTimeout));